Add basic tests, ci, + make mathbox mount its own container
This adds some basic tests and CI using jest.

It may be worth considering karma + jasmine instead of jest. I prefer jest, but I think there is value in NOT mocking MathBox, since if this library fails to integrate with mathbox, then this library is useless. Mathbox really wants a functional browser with WebGL contexts which Jest does not provide.

To circumvent this, I've added an "auto-mock" using proxies which auto-vivifies properties and methods similar to python Mock objects. This is cool and is working, but is sufficiently non-standard (for ts/js) that it makes me slightly nervous.