fix rerendering errors (#20)
This commit fixes issues with children rendeirng inside outdated parents. Two things:

First: when parents change, destroy the old children and create new ones.  We weren't doing that before.

Second: Check that the parent-according-to-react is the same as the parent-according-to-mathbox. If they are different, destroy-and-create as above...

But...why might they be different? I believe fundamentally the issue is:
- we store parent value in context...
- we update the context parent value using useEffect hooks..
- useEffect hooks are called children before parents.

One wonders whether we really need to be using useEffect hooks for the mathbox components, or if we can just put all the node creation/removal stuff in the render function. Hooks seem nice for dependencies, but mathbox is pretty smart about diffing out the actual changes.