only allow some components to have children (#27)
Because the MathBox API is "fluid", i.e., designed around chaining, node creation can always be chained:

```js
const root = mathBox()

root
  .cartesian()  // create cartesian... parent = root
  .volume()     // create volume...    parent = cartesian
  .point()      // create point...     parent = cartesian
```

The `.point()` call might appear as if its parent would be volume, but in actuality its parent is cartesian.

Previously, mathbox-react allowed all components to have children. So this was allowed:

```
<Cartesian>
  <Volume>
    <Points />
  </Volume>
</Cartesian>
```
In order to ensure that the mathbox tree and react tree stay in-sync, mathbox-react checks that the react-node's parent and the mathbox-node parent are the same and re-renders if they are not. But the above component tree never gets in-sync, because MathBox always renders Point as a child of Cartesian, never as a child of Volume.

With this commit, mathbox-react becomes a bit more strict about children. Now, a component can only have children if mathbox will actually place the children on that node.

Resolves https://github.com/ChristopherChudzicki/mathbox-react/issues/25